refactored int ids to long

+++ src/main/java/com/ziggs/ziggs_backend/entity/Alert.java
@@ -13,7 +13,7 @@ public class Alert {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "alert_id")
-    private Integer alertId;
+    private Long alertId;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
